requirement.txt
djangorestframework
pygments
markdown
bootstrap-admin
